chore: move `Color.kt`, `Type.kt` and `Dimens.kt`
- Moved to `app/src/main/java/com/android/periodpals/resources` for code coverage (?)
