+++ app/src/main/java/com/android/periodpals/ui/theme/AppUtil.kt
@@ -4,6 +4,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.compositionLocalOf
 import androidx.compose.runtime.remember
+import com.android.periodpals.resources.CompactLargeDimens
+import com.android.periodpals.resources.Dimens
 
 /**
  * A composable function that provides the application dimensions to the composition.
